implement using transform callback(ish)
