+++ test/context_through_slot_test.exs
@@ -12,7 +12,7 @@ defmodule Surface.ContextThroughSlotTest do
 
     def render(assigns) do
       ~H"""
-      <Context :set={{ :foo, @foo }}>
+      <Context set={{ :foo, @foo }}>
         <slot />
       </Context>
       """
@@ -41,7 +41,7 @@ defmodule Surface.ContextThroughSlotTest do
     def render(assigns) do
       # @foo is nil here
       ~H"""
-      <Context :get={{ :foo }}>
+      <Context get={{ :foo }}>
         <div>{{ @foo }}</div>
       </Context>
       """
